Home

Photon

Photon is a developer friendly framework for Deep Learning in Julia. Under the hood it leverages Knet and it provides a Keras like API on top of that.

Steps

To train your own model, there are three steps to follow:

Create your model using the layers that come out of the box with Photon or using your own custom layers.

Create a Workout that combines the model, a loss function and an optimiser. Optionally you can also add some metrics that you want to monitor.

Train the model by calling fit! on the workout and the training data.

Step 1: Model

A model can use out of the box layers or your own layers. Photon support most common type of layer:

Dense or fully connected layers.

2D ad 3D convolutional layers

Different type of Pooling layers

Recurrent layers (RNN, LSTM, GRU)

Dropout layers.

Several container type layers like Sequential and Residual

Some examples how to create the different type of models:

mymodel = Sequential(
            Dense(64, relu),
            Dense(10)
)

mymodel = Sequential(
            Conv2D(64, 3, relu),
            Conv2D(64, 3, relu),
            MaxPool(),
            Dense(10)
)

mymodel = Sequential(
            LSTM(64),
            Dense(64),
            Dense(10)
)

So normally you won't need to create your own layers. But if you have to, a layer is nothing more than function. So the following could be a layer ;)

myLayer(X) = moon == :full ? X .- 1 : X

Step 2: Workout

A workout combines a model + loss + optimiser and keeps track of the progress during the actual training. The workout is stateful in the sense that you can run multiple training sessions and the progress will be recorded appropriately.

The minimum required to create a workout is:

workout = Workout(mymodel, MSE())

Besides this, you can pass an optimzer and define the metrics that you want to get tracked during the training sessions. Photon tracks :loss and :val_loss (for the validation phase) by default, but you define additional ones.

workout = Workout(mymodel, MSE(), SGD(), acc=accuracy())

Step 3: fit!

The actual training is done using the fit! function.

fit!(workout, data, epochs=5)

If you provide also data for the validation phase, Photon will automatically run a validation after a training epoch has finished.

fit!(workout, data, training_data, epochs=10)

The data is expected to be a tuple of (X, Y) where X and Y can be tuples again in case your model expects multiple inputs or outputs. So some examples of valid formats

(X,Y)
((X2,X2), Y)
(X, (Y1, Y2, Y3))
((X1, X2), (Y1, Y2))

By default fit! will convert each batch to the right data type and device. This is controlled by the optional parameter convertor. If you don't want a convertor and ensure the provided data is already in the right format, you can pass the identity function:

fit!(workout, data; convertor=identity)

API

Photon.GRU — Function.

Create a GRU layer

Examples:

layer = GRU(50)

source

Photon.KorA — Method.

KorA makes it easy to move an array to the GPU or the other way around

source

Photon.LSTM — Function.

Create a LSTM layer.

Examples:

layer = LSTM(50)

source

Photon.RNN — Function.

A simple RNN layer.

source

Photon.fit! — Function.

Train the model based on a supervised dataset and for a number of epochs. fit! can be called multiple times and will continue to train where is left of last time.

By default the fit! function will try to ensure the data is of the right type (e.g. Float32) and on the right device (e.g. GPU) before feeding it to the model.

Usage

fit!(workout, traindata)
fit!(workout, traindata, testdata, epochs=50)

If you don't want any data conversion, just pass the identity funciton as the convertor:

fit!(workout, traindata, convertor=identity)

source

Photon.hasmetric — Method.

Does the workout have any recorded values for a certain metric

source

Photon.loadWorkout — Method.

Load a workout from file and return it.

Usage

workout = loadWorkout("workout_1000.dat")
fit!(workout, mydata)

source

Photon.output_size — Method.

Utility function to determine the output size of a convolutional layer given a certain input size and configuration of a convolutional layer. This function works even if the weights of the layer are not initialized.

For example: c = Conv2D(16, (3,3), strides=3, padding=1) output_size(c, (224,224)) # output is (75, 75)

source

Photon.plotmetrics — Function.

Plot the metrics after some training. This function will plot all the metrics in a single graph.

In order to avoid Photon being dependend on Plots, the calling code will have to provide that module as the first parameter.

Usage

fit!(workout, mydata, epochs=10)

import Plots
plotmetrics(Plots, workout)

source

Photon.predict — Method.

Predict a sample, either a single value or a batch. Compared to invoking the model directory with model(x), predit takes care of:

Moving the data to the GPU if required.

Making the data into a batch (controlled by makebatch parameter)

Usage

x = randn(Float32, 224, 224, 3)
predict(model, x)

source

Photon.saveWorkout — Function.

Save a workout to a file. This will save all the state that is captured in the workout and enables to continue at a later stage.

source

Photon.AdaptiveAvgPool — Type.

Adaptive Average Pool has a fixed size output and enables creating a convolutional network that can be used for multiple image formats.

source

Photon.AdaptiveMaxPool — Type.

Adaptive MaxPool has a fixed size output and enables creating a convolutional network that can be used for multiple image formats.

source

Photon.BCELoss — Type.

Binary CrossEntropy

source

Photon.BatchNorm — Type.

BatchNorm layer with support for an optional activation function

source

Photon.Concurrent — Type.

Concurrrent

source

Photon.ContextSwitch — Type.

Beginning of allowing for a single model instance to run on multiple devices (expiremental)

source

Photon.CrossEntropyLoss — Type.

CrossEntropy loss function with support for an optional weight parameter. The weight parameter can be static (for example to handle class inbalances) or dynamic (so passed every time when the lost function is invoked)

Usage

workout = Workout(model, CE(), SGD())

source

Photon.Dense — Type.

Fully connected layer with an optional bias weight.

Usage

layer = Dense(10, relu)
layer = Dense(100, use_bias=false)

source

Photon.Dropout — Type.

Dropout layer with optional the rate (between 0 and 1) of dropout. If no rate is specified, 0.5 (so 50%) will be used.

source

Photon.Flatten — Type.

Flattening Layer. Photon by default already has flattening funcitonality build into the Dense layer, so you won't need to include a separate Flatten layer before a Dense layer.

source

Photon.HingeLoss — Type.

Hinge Loss implementation

source

Photon.MAELoss — Type.

Mean Absolute Error implementation, also referred to as the L1 Loss.

source

Photon.MSELoss — Type.

Mean Square Error implementation, also referred to as the L2 Loss.

Usage

workout = Workout(model, MSELoss(), SGD())

source

Photon.PseudoHuberLoss — Type.

Pseudo Huber Loss implementation, somewhere between a L1 and L2 loss.

source

Photon.Residual — Type.

Residual Layer. This will stack on the second last dimension. So with and 2D convolution this will be the channel layer (WxHxCxN)

source

Photon.Sequential — Type.

Sequential

source

Photon.Workout — Type.

The Workout keeps track of the progress of the training session. At least a model and a loss function needs to be provided. Optional an optimizer and one or more metrics can be specified.

If no optimizer is provided, SGD will be used. If no metrics are provided, only the loss during training and validation will be registered (:loss and :val_loss).

Usage

workout = Workout(model, mse)

workout = Workout(model, nll, SGD())

workout = Workout(model, nll, SGD(); acc=BinaryAccuracy())

source

Photon.autoConvertor — Method.

autoConvertor converts data to the right format for a model. It uses the context to determine the device (cpu or gpu) and datatype that the data needs to be.

It supports Tuples, Arrays and KnetArrays and a combination of those.

source

Photon.back! — Method.

Perform the back propagation and update of weights in one go.

source

Photon.getmetricname — Function.

Function to generate the fully qualified metric name. It uses the metric name and phase (train or valid) to come up with a unique name.

source

Photon.getmetricvalue — Function.

Get the metric value for a fully qualified metric name and a certain step. If step is not provided the last step is used. If no value is found the passed function will not be invoked.

Usage

getmetricvalue(workout, :val_loss) do value
    println("validation loss", value)
end

source

Photon.gradients — Function.

Utility function to calculate the gradients. Useful when checking for vanishing or exploding gradients. The returned value is a Vector of (Param, Gradient).

source

Photon.step! — Method.

Take a single step in updating the weights of a model. This function will be invoked from fit! to do the actual learning.

For a minibatch (x,y) of data, the folowing sequence will be executed:

perform the forward pass

calculate the loss

update and remember the metrics, if any

do the backpropagation and update the weights

source

Photon.updatemetrics! — Function.

Invoke the configured metrics. The loss metric will always be logged and available. Metrics are stored in the history attribute of the workout.

source

Photon.validate — Method.

Validate a minibatch and calculate the loss and metrics. Typically this function is called from the fit! method.

source

Serialization.serialize — Method.

Enable saving and loading of models by specialized KnetArray methods for Julia serialization This will effectively move a GPU weight to the CPU before serialing it and move it back to the GPU when deserializing.

source

Photon.Context — Type.

Context is used by various parts of Photon to determine what the device and datatype should be for Arrays. It also allows to quickly switch between GPU and CPU based models.

Attributes

device::Symbol the type of device. For now supported :cpu and :gpu

deviceId::Int the id of the device, useful for example if you multiple GPU's

dtype::Type the type of data you want to use. Most common are Float32, Float16 or Float64.

source

Photon.Conv — Type.

Convolutional layer that serves as the base for Conv2D and Conv3D

source

Photon.ConvTranspose — Type.

ConvTranspose layer

source

Photon.Loss — Type.

Base type for the loss functions. However Photon accepts any functoon as a loss function as long as it is callable and returns the loss as a scalar value.

 fn(y_pred, y_true) :: Number

source

Photon.Meter — Type.

A meter is reponsible for presenting metric values. This can be printing it to the console output, showing it on TensorBoard of storing it in a database.

source

Photon.PoolingLayer — Type.

Pooling layers

source

Photon.StackedLayer — Type.

Common behavior for stacked layers that enables to access them as arrays

source