Core

Basic

Photon.Workout — Type.

The Workout keeps track of the progress of the training session. At least a model and a loss function needs to be provided. Optional an optimizer and one or more metrics can be specified.

If no optimizer is provided, SGD will be used. If no metrics are provided, only the loss during training and validation will be registered (:loss and :val_loss).

Usage

workout = Workout(model, L1Loss())

workout = Workout(model, CrossEntropy(), SGD())

workout = Workout(model, HingeLoss(), SGD(); acc=BinaryAccuracy())

source

Photon.saveWorkout — Function.

Save a workout to a file. This will save all the state that is captured in the workout and enables to continue at a later stage.

source

Photon.loadWorkout — Function.

Load a workout from file and return it.

Usage

workout = loadWorkout("workout_1000.dat")
fit!(workout, mydata)

source

Photon.validate — Function.

Validate a minibatch and calculate the loss and metrics. Typically this function is called from the fit! method. But if required can also be invoked directly.

source

Photon.predict — Function.

Predict a sample, either a single value or a batch. Compared to invoking the model directory with model(x), predit takes care of:

Moving the data to the GPU if required.

Making the data into a batch (controlled by makebatch parameter)

Usage

x = randn(Float32, 224, 224, 3)
predict(model, x)

source

Photon.fit! — Function.

Train the model based on a supervised dataset and for a number of epochs. fit! can be called multiple times and will continue to train where is left of last time.

By default the fit! function will try to ensure the data is of the right type (e.g. Float32) and on the right device (e.g. GPU) before feeding it to the model.

Usage

fit!(workout, traindata)
fit!(workout, traindata, testdata, epochs=50)

If you don't want any data conversion, just pass the identity funciton as the convertor:

fit!(workout, traindata, convertor=identity)

source

Photon.Context — Type.

Context is used by various parts of Photon to determine what the device and datatype should be for data. It allows for quickly switches between GPU and CPU based environments.

Attributes

device::Symbol the type of device. For now supported :cpu and :gpu

deviceId::Int the id of the device, useful for example if you multiple GPU's

dtype::Type the type of data you want to use. Most common are Float32, Float16 or Float64.

source

Photon.freeze! — Function.

Freeze a parameter so it no longer will be updated during training.

source

Photon.unfreeze! — Function.

Unfreeze a parameter so it will be updated again during training.

source

Internal

You normally won't have to invoke the following functions directly when training a model. But in some cases you might want to write a specialised version of them.

Photon.back! — Function.

Perform the back propagation and update of weights in a single go.

source

Photon.step! — Function.

Take a single step in updating the weights of a model. This function will be invoked from fit! to do the actual learning.

For a minibatch (x,y) of data, the folowing sequence will be executed:

perform the forward pass

calculate the loss

update and remember the metrics, if any

do the backpropagation and update the weights

source