Layers

Basic

Photon.Dense — Type.

Regular densely-connected NN layer.

Dense implements the operation: output = activation(dot(input, weight) + bias) where activation is the element-wise activation function passed as the activation argument, weight is a weights matrix created by the layer, and bias is a bias vector created by the layer (only applicable if use_bias is true).

Usage

layer = Dense(10, relu)
layer = Dense(100, use_bias=false)

source

Photon.Dropout — Type.

Dropout layer with optional the rate (between 0 and 1) of dropout. If no rate is specified, 0.5 (so 50%) will be used.

source

Photon.BatchNorm — Type.

Batch normalization layer (Ioffe and Szegedy, 2014). Normalizes the input at each batch, i.e. applies a transformation that maintains the mean activation close to 0 and the activation standard deviation close to 1.

Finally, if activation is not nothing, it is applied to the outputs as well.

source

Photon.Flatten — Type.

Flattening Layer. Photon by default already has flattening funcitonality build into the Dense layer, so you won't need to include a separate Flatten layer before a Dense layer.

source

Container

Photon.Sequential — Type.

Sequential layer allows to chain together a number of other layers.

Usage

model = Sequential(Conv2D(100),MaxPool(),Dense(10))

source

Photon.Concurrent — Type.

Concurrrent layer allows for stacking a number of other layers in parallel and combining their results before returning it.

This layer will stack on the second last dimension. So with 2D and 3D convolution this will be the channel layer (WxHxCxN). As a result other dimensions have to the same.

source

Photon.Residual — Type.

Residual Layer works like a Sequential layer, however before returning the result it will be combined with the orginal input (residual). This is a popular techique in modern neural networds since it allows for better backpropagation.

This will stack on the second last dimension. So with 2D and 3D convolution this will be the channel layer (WxHxCxN)

source

Convolutional

Photon.Conv2D — Type.

2D convolution layer (e.g. spatial convolution over images).

This layer creates a convolution kernel that is convolved with the layer input to produce a tensor of outputs. If use_bias is true, a bias vector is created and added to the outputs. Finally, if activation is not nothing, it is applied to the outputs as well.

source

Photon.Conv2DTranspose — Type.

ConvTranspose layer

source

Photon.Conv3D — Type.

3D convolution layer (e.g. spatial convolution over volumes).

This layer creates a convolution kernel that is convolved with the layer input to produce a tensor of outputs. If use_bias is true, a bias vector is created and added to the outputs. Finally, if activation is not nothing, it is applied to the outputs as well.

source

Photon.Conv3DTranspose — Type.

ConvTranspose layer

source

Pooling

Missing docstring.

Missing docstring for MaxPool2D. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AvgPool2D. Check Documenter's build log for details.

Missing docstring.

Missing docstring for MaxPool3D. Check Documenter's build log for details.

Missing docstring.

Missing docstring for AvgPool3D. Check Documenter's build log for details.

Photon.AdaptiveAvgPool — Type.

Adaptive Average Pool has a fixed size output and enables creating a convolutional network that can be used for multiple image formats.

source

Photon.AdaptiveMaxPool — Type.

Adaptive MaxPool has a fixed size output and enables creating a convolutional network that can be used for multiple image formats.

source

Recurrent

Photon.RNN — Function.

A simple RNN layer.

source

Photon.LSTM — Function.

Applies a multi-layer long short-term memory (LSTM) RNN to an input sequence.

For each element in the input sequence, each layer computes the following function:

$\begin{array}{ll} i_t = sigmoid(W_{ii} x_t + b_{ii} + W_{hi} h_{(t-1)} + b_{hi}) \\f_t = sigmoid(W_{if} x_t + b_{if} + W_{hf} h_{(t-1)} + b_{hf}) \\g_t = \tanh(W_{ig} x_t + b_{ig} + W_{hc} h_{(t-1)} + b_{hg}) \\o_t = sigmoid(W_{io} x_t + b_{io} + W_{ho} h_{(t-1)} + b_{ho}) \\c_t = f_t * c_{(t-1)} + i_t * g_t \\h_t = o_t * \tanh(c_t) \end{array}$

Examples:

layer = LSTM(50)
layer = LSTM(50, 2, bidirectional=true)

source

Photon.GRU — Function.

Create a GRU layer

Examples:

layer = GRU(50)

source