Losses

Loss functions

Loss functions can be used both to calculate the loss as the last step in a neural network as well as calculate metrics. In either cases they are provided as an argument to the Workout constructor

# L1Loss as a loss
workout = Workout(model, L1Loss())

# BCELoss as a loss function and FocalLoss as a metric
workout = Workout(model, BCELoss(), floss=FocalLoss())

Photon.Losses.Loss — Type.

Abstract type for the loss functions. However Photon accepts any function as a loss function as long as it is callable and returns the loss value as a scalar type.

source

Missing docstring.

Missing docstring for L1Loss. Check Documenter's build log for details.

Missing docstring.

Missing docstring for L2Loss. Check Documenter's build log for details.

Missing docstring.

Missing docstring for LNLoss. Check Documenter's build log for details.

Missing docstring.

Missing docstring for PseudoHuberLoss. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BCELoss. Check Documenter's build log for details.

Missing docstring.

Missing docstring for CrossEntropyLoss. Check Documenter's build log for details.

Missing docstring.

Missing docstring for HingeLoss. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SquaredHingeLoss. Check Documenter's build log for details.

Missing docstring.

Missing docstring for FocalLoss. Check Documenter's build log for details.