Core

Basic

Photon.Workout — Type.

The Workout keeps track of the progress of the training session. At least a model and a loss function needs to be provided. Optional an optimizer and one or more metrics can be specified.

If no optimizer is specified, SGD will be used. If no metrics are provided, only the loss during training and validation will be registered (:loss and :val_loss).

The provided mover will move data to the correct device. See also SmartMover. If no mover is required, you can provide the identity function instead.

Usage

workout = Workout(model, L1Loss())

workout = Workout(model, CrossEntropy(), Adam())

workout = Workout(model, HingeLoss(); acc=BinaryAccuracy())

workout = Workout(model, L1Loss(), mover=identity)

source

Missing docstring.

Missing docstring for saveWorkout. Check Documenter's build log for details.

Missing docstring.

Missing docstring for loadWorkout. Check Documenter's build log for details.

Photon.validate — Function.

Validate a minibatch and calculate the loss and metrics. Typically this function is called from the train! method. But if required can also be invoked directly.

source

Photon.predict — Function.

Predict a sample, either a single value or a batch. Compared to invoking the model directory with model(x), predit takes care of:

Moving the data to the GPU if required.

Shaping the data into a batch (controlled by makebatch parameter)

Usage

x = randn(Float32, 224, 224, 3)
predict(workout, x)

source

Photon.train! — Function.

Train the model based on a supervised dataset and for a number of epochs. train! can be called multiple times and will continue to train where is left of last time.

By default the train! function will try to ensure the data is of the right type (e.g. Float32) and on the right device (e.g. GPU) before feeding it to the model.

Usage

train!(workout, traindata)
train!(workout, traindata, testdata, epochs=50)

source

Photon.freeze! — Function.

Freeze a parameter so it no longer will be updated during training.

source

Photon.unfreeze! — Function.

Unfreeze a parameter so it will be updated again during training.

source

Photon.stop — Function.

Stop a training session. Typically invoked by a callback function that detects that the training is not progressing anymore.

If this function is called outside the scope of a trianing session, an exception is thrown.

source

Photon.gradients — Function.

Utility function to calculate the gradients. Useful when checking for vanishing or exploding gradients. The returned value is a Vector of (Param, Gradient).

source

Context

Context is a core concept in Photon. It determines what device (GPU or CPU) to use and what datatype (Float64/32/16) is required. Once the context is set, models and data will use this by default.

Photon.Context — Type.

Context is used by various parts of Photon to determine what the device and datatype should be for data. It allows for quickly switches between GPU and CPU based environments.

Attributes

device::Symbol the type of device. For now supported :cpu and :gpu

deviceId::Int the id of the device, useful for example if you multiple GPU's

dtype::Type the type you want to use for parameters. Most common are Float32, Float16 or Float64.

source

Missing docstring.

Missing docstring for getContext. Check Documenter's build log for details.

Missing docstring.

Missing docstring for setContext. Check Documenter's build log for details.

Missing docstring.

Missing docstring for resetContext. Check Documenter's build log for details.

Photon.Mover — Type.

Moves data to the right device like a CPU or GPU. However implememetations can provide extra functionality like also taking care of the correct data types.

The default Mover used by Photon is SmartMover if no other mover is specified during instantiation of the Workout.

source

Photon.SmartMover — Type.

SmartMover converts data to the right device and optional data type for a model. It uses the context to determine the device (cpu or gpu) and datatype that the data needs to be.

It move_float is true, SmartMover will ensure that any provided Array of the type AbstractFloat will convert to the Float type as defined in the context.

It supports Tuples, Arrays and KnetArrays and a combination of those.

source

Internal

You normally won't have to invoke the following functions directly when training a model. But in some cases you might want to write a specialized version of them.

Photon.back! — Function.

Perform the back propagation and update of weights in a single go.

source

Photon.step! — Function.

Take a single step in updating the weights of a model. This function will be invoked from train! to do the actual learning.

For a minibatch (x,y) of data, the folowing sequence will be executed:

perform the forward pass

calculate the loss

update and remember the metrics, if any

do the backpropagation and update the weights

source