Data

Datasets

Photon.Data.Dataset — Type.

Datasets can load and transform data and can be used in a pipeline. Typically they would have length and index methods implemented and need to return only a single sample. The MiniBatch transformer would collect these single samples and put them into a mini batch.

data = SomeDataset() |> SomeTransformer() |> MiniBatch()
train!(workout, data)

Please note that in case all your data would fit in memory, you can feed it directly to the train! function (so a dataset is not required)

X = [randn(10,10,16) for i in 1:100]
Y = [randn(1,16) for i in 1:100]
train!(workout, zip(X,Y))

source

Photon.Data.ImageDataset — Type.

Dataset that loads an single image from a file and optionally resizes the image. The labels are passed as is.

Usage

ds = ImageDataset(filenames, labels, resize=(200,200))

source

Photon.Data.JLDDataset — Type.

Dataset that loads data from a JLD(2) file. The format that is stored is expected to be in the shape of: (X,Y)

Note: Tested with JLD2 library only.

Example:

using JLD2

jldopen("example.jld2", "w") do file
    file["image1"] = (randn(Float32,28,28,1), rand(0:9,1))
    file["image2"] = (randn(Float32,28,28,1), rand(0:9,1))
end

f = jldopen("example.jld2", "r")
ds = JLD2Dataset(f)
ds[1]

source

Photon.Data.DFDataset — Type.

Dataset that retrieves is data from a dataframe. The provided column names for X and Y can be either a single Symbol or a Vector of Symbols.

Usage

df = DataFrame(randn(4,20))
ds = DFDataset(df, :x1, :x2)
ds = DFDataset(df, [:x1, :x3], [:x2,:x4])

source

Missing docstring.

Missing docstring for JuliaDBDataset. Check Documenter's build log for details.

Photon.Data.VectorDataset — Type.

Dataset that contains two vectors, one for the input data and one for the labels.

source

Photon.Data.TestDataset — Type.

A dataset that contains random generated samples. Ideal for quick testing of a model. The random values will be drawn from a normal distribution.

You need to provide the shape of X, Y and the number of samples that the dataset should contain. Optionally you can specify a sleep value to simulate IO blocking.

Usage

xshape = (28,28,1)
yshape = (10,)
ds = TestDataset(xshape, yshape, 60000)

ds = TestDataset((100,), (1,), 1000, sleep=0.1)

source

Transformers

Photon.Data.Transformer — Type.

Transformers enable you to transform the output generated by a datasets. Many of the transformers can be chained together.

data = SomeDataset(...)
data = data |> Transformer1(...) |> Transformer2(...) |> Transformer3(...)

train!(workout, data)

source

Photon.Data.ImageCrop — Type.

Randomly crop an image in the shape of WxHxC to a certain size. This Transformer uses plain Array operations and doesn't rely on packages like Images.jl

Usage

# Only crop X
ds = TestDataset((100,100,3),(10,),100)
ds = ds |> ImageCrop((50,50))

# Crop both X and Y
ds = TestDataset((100,100,3),(50,50,3),100)
ds = ds |> ImageCrop((50,50), (20,20))

source

Photon.Data.Normalizer — Type.

Normalize data before feeding it to a model. A typical usecase is to normalize an image. By default will only normalize to the input data (:X) and not the labels (:Y).

Usage

ds = TestDataset((28,28,1),(10,),100) |> Normalize()

source

Photon.Data.OneHotEncoder — Type.

OneHot transformer for a single sample. Uses the onehot function to perform the actual transoformation.

Supports both single and multi class encoding:

5 => [0,0,0,0,0,1,0,0,0,0]
[5,7] => [0,0,0,0,0,1,0,1,0,0]

Usage

ds = mnistds() |> OneHotEncoder(0:9)

source

Photon.Data.Subset — Type.

Return a subset of samples of the underlying dataset. Optionally you can specify a start index (default = 1).

Usage

data = MyDataset() |> Subset(1:100)

source

Photon.Data.NoisingTransfomer — Type.

Add noise to the samples in a dataset. By default will only add noise to the input data (:X) and not the labels (:Y).

Usage

ds = TestDataset((28,28,1),(10,),100)
ds = ds |> NoiseTransformer(noiselevel=0.05)

source

Photon.Data.MiniBatch — Type.

A MiniBatch is a transformer that will collect multiple samples from the underlying dataset and put it in a minibatch. So it will call a dataset to get a single sample and combines them into a minibatch. The MiniBatch itself is an iterator.

It will do this using threading, so when a dataset will read some samples from disk this won't become a botttleneck.

Usage

data = ImageDataset(filenames,labels) |> MiniBatch(16, shuffle=False)
train!(workout, data)

source

Photon.Data.Split — Type.

Split one dataset into two subsets, one for training and one for validation. By default Split will shuffle the indexes first in order to get a representative validaiton set. The default split is 80% for training and 20% for validation.

Usage

data_train, data_valid = ImageDataset(...) |> Split(0.25, shuffle=false)

source