Layers

Basic

Photon.Dense — Type.

Regular densely-connected NN layer.

Dense implements the operation: output = activation(dot(input, weight) + bias) where activation is the element-wise activation function passed as the activation argument, weight is a weights matrix created by the layer, and bias is a bias vector created by the layer (only applicable if use_bias is true).

Usage

layer = Dense(10, relu)
layer = Dense(100, use_bias=false)

source

Photon.Dropout — Type.

Dropout layer with optional the rate (between 0 and 1) of dropout. If no rate is specified, 0.5 (so 50%) will be used.

source

Photon.BatchNorm — Type.

Batch normalization layer (Ioffe and Szegedy, 2014). Normalizes the input at each batch, i.e. applies a transformation that maintains the mean activation close to 0 and the activation standard deviation close to 1.

Finally, if activation is not nothing, it is applied to the outputs as well.

source

Photon.Flatten — Type.

Flattening Layer. Photon by default already has flattening funcitonality build into the Dense layer, so you won't need to include a separate Flatten layer before a Dense layer.

source

Container

Containers are special type of layers that contain other layers. They typically extend the abstract type that allows to use them as regular Vectors.

Photon.StackedLayer — Type.

Stacked layer is a abstract type and defines common behavior for containers layers that enables to access them as arrays.

source

Photon.Sequential — Type.

Sequential layer allows to chain together a number of other layers.

Usage

model = Sequential(Conv2D(100),MaxPool(),Dense(10))

source

Photon.Concurrent — Type.

Concurrrent layer allows for stacking a number of other layers in parallel and combining their results before returning it.

This layer will stack on the second last dimension. So with 2D and 3D convolution this will be the channel layer (WxHxCxN). As a result other dimensions have to the same.

source

Photon.Residual — Type.

Residual Layer works like a Sequential layer, however before returning the result it will be combined with the orginal input (residual). This is a popular techique in modern neural networds since it allows for better backpropagation.

This will stack on the second last dimension. So with 2D and 3D convolution this will be the channel layer (WxHxCxN)

source

Convolutional

Photon contains convolutional layers for 2D and 3D spatial data.

A 2D convolutional layer would require a 4D Array in the shape of WxHxCxN (width x height x channels x batch). So for a typical image classification problem this could look like: 224 x 224 x 3 x 8 (224 by 224 image, with 3 colors and 8 samples per batch).

Photon.Conv2D — Type.

2D convolution layer (e.g. spatial convolution over images).

This layer creates a convolution kernel that is convolved with the layer input to produce a tensor of outputs. If use_bias is true, a bias vector is created and added to the outputs. Finally, if activation is not nothing, it is applied to the outputs as well.

source

Photon.Conv2DTranspose — Type.

Transposed 2D convolution layer (sometimes called Deconvolution).

The need for transposed convolutions generally arises from the desire to use a transformation going in the opposite direction of a normal convolution, i.e., from something that has the shape of the output of some convolution to something that has the shape of its input while maintaining a connectivity pattern that is compatible with said convolution.

source

Photon.Conv3D — Type.

3D convolution layer (e.g. spatial convolution over volumes).

This layer creates a convolution kernel that is convolved with the layer input to produce a tensor of outputs. If use_bias is true, a bias vector is created and added to the outputs. Finally, if activation is not nothing, it is applied to the outputs as well.

source

Photon.Conv3DTranspose — Type.

Transposed 3D convolution layer (sometimes called Deconvolution).

The need for transposed convolutions generally arises from the desire to use a transformation going in the opposite direction of a normal convolution, i.e., from something that has the shape of the output of some convolution to something that has the shape of its input while maintaining a connectivity pattern that is compatible with said convolution.

source

Pooling

Photon.PoolingLayer — Type.

Pooling layers

source

Photon.MaxPool2D — Type.

Max pooling operation for two dimensional (spatial) data.

source

Photon.AvgPool2D — Type.

Average pooling operation for two dimensional (spatial) data.

source

Photon.MaxPool3D — Type.

Max pooling operation for three dimensional (spatial) data.

source

Photon.AvgPool3D — Type.

Average pooling operation for three dimensional (spatial) data.

source

Photon.AdaptiveAvgPool — Type.

Adaptive Average Pool has a fixed size output and enables creating a convolutional network that can be used for multiple image formats.

source

Photon.AdaptiveMaxPool — Type.

Adaptive MaxPool has a fixed size output and enables creating a convolutional network that can be used for different image sizes.

source

Recurrent

Photon.RNN — Function.

Applies a multi-layer Elman RNN with :tanh or :relu non-linearity to an input sequence.

For each element in the input sequence, each layer computes the following function:

$h_t = \text{tanh}(W_{ih} x_t + b_{ih} + W_{hh} h_{(t-1)} + b_{hh})$

source

Photon.LSTM — Function.

Applies a multi-layer long short-term memory (LSTM) RNN to an input sequence.

For each element in the input sequence, each layer computes the following function:

$\begin{array}{ll} i_t = sigmoid(W_{ii} x_t + b_{ii} + W_{hi} h_{(t-1)} + b_{hi}) \\f_t = sigmoid(W_{if} x_t + b_{if} + W_{hf} h_{(t-1)} + b_{hf}) \\g_t = \tanh(W_{ig} x_t + b_{ig} + W_{hc} h_{(t-1)} + b_{hg}) \\o_t = sigmoid(W_{io} x_t + b_{io} + W_{ho} h_{(t-1)} + b_{ho}) \\c_t = f_t * c_{(t-1)} + i_t * g_t \\h_t = o_t * \tanh(c_t) \end{array}$

Usage

layer = LSTM(50)
layer = LSTM(50, 2, bidirectional=true)

source

Photon.GRU — Function.

Applies a multi-layer gated recurrent unit (GRU) RNN to an input sequence.

For each element in the input sequence, each layer computes the following function:

$\begin{array}{ll} r_t = \sigma(W_{ir} x_t + b_{ir} + W_{hr} h_{(t-1)} + b_{hr}) \\z_t = \sigma(W_{iz} x_t + b_{iz} + W_{hz} h_{(t-1)} + b_{hz}) \\n_t = \tanh(W_{in} x_t + b_{in} + r_t * (W_{hn} h_{(t-1)}+ b_{hn})) \\h_t = (1 - z_t) * n_t + z_t * h_{(t-1)} \end{array}$

Usage

layer = GRU(50)
layer = GRU(50, 2, bidirectional=true)

source

Experimental

Photon.ContextSwitch — Type.

Beginning of allowing for a single model instance to run on multiple devices (expiremental)

source