Losses

Loss functions

Loss functions can be used both to calculate the loss as the last step in a neural network as well as calculate metrics. In either cases they are provided as an argument to the Workout constructor

# L1Loss as a loss
workout = Workout(model, L1Loss())

# BCELoss as a loss function and FocalLoss as a metric
workout = Workout(model, BCELoss(), floss=FocalLoss())

Photon.Loss — Type.

Abstract type for the loss functions. However Photon accepts any function as a loss function as long as it is callable and returns the loss value as a scalar type.

source

Photon.L1Loss — Type.

Calculates the mean absolute error between label and pred.

$L = \sum_i \vert {label}_i - {pred}_i \vert$

label and pred can have arbitrary shape as long as they have the same number of elements.

source

Photon.L2Loss — Type.

Calculates the mean squared error between label and pred.

$L = \frac{1}{2} \sum_i ({label}_i - {pred}_i)^2$

label and pred can have arbitrary shape as long as they have the same number of elements.

source

Photon.LNLoss — Type.

More generic version of L1loss and L2loss that allows L{N}Loss.

$L = \frac{1}{2} \sum_i \vert {label}_i - {pred}_i \vert^{N}$

label and pred can have arbitrary shape as long as they have the same number of elements.

source

Photon.PseudoHuberLoss — Type.

Pseudo Huber Loss implementation, somewhere between a L1 and L2 loss.

source

Photon.BCELoss — Type.

BCE (Binary Cross Entropy) loss is useful when training logistic regression.

$L = - \sum_i {label}_i * \log({pred}_i) + (1 - {label}_i) * \log(1 - {pred}_i)$

If use_sigmoid is true, first a sigmoid activation will be applied before calcuating the BCE loss.

source

Photon.CrossEntropyLoss — Type.

CrossEntropy loss function with support for an optional weight parameter. The weight parameter can be static (for example to handle class inbalances) or dynamic, so passed every time when the lost function is invoked.

Cross Entropy Function:

$L = - \sum_i {label}_i \log({output}_i)$

If use_softmax is true, first the softmax(output) will be performed before cross entropy.

Usage

workout = Workout(model, CrossEntropyLoss(), SGD())

source

Photon.HingeLoss — Type.

Calculates the hinge loss function often used in SVMs:

$L = \sum_i max(0, {margin} - {pred}_i \cdot {label}_i)$

where pred is the classifier prediction and label is the target tensor containing values -1 or 1. label and pred must have the same number of elements.

If autofix is true, will convert label from {0,1} to {-1,1}

source

Photon.SquaredHingeLoss — Type.

Calculates the soft-margin loss function used in SVMs:

$L = \sum_i max(0, {margin} - {pred}_i \cdot {label}_i)^2$

where pred is the classifier prediction and label is the target tensor containing values -1 or 1. label and pred can have arbitrary shape as long as they have the same number of elements.

source

Photon.FocalLoss — Type.

Focal loss for multi-classification

$L=-alpha(1-p_t)^{gamma}ln(p_t)$

See also Focal Loss for Dense Object Detection https://arxiv.org/abs/1708.02002

source