Metrics

Metrics

Metrics can be passed to the Workout initializer and provide extra insights next to the loss value. Every metric is invoked with two parameters: predicted outcome and the labels.

Photon.Metrics.BinaryAccuracy — Type.

Binary accuracy calculation

source

Photon.Metrics.OneHotBinaryAccuracy — Type.

Binary accuracy for a onehot classification.

source

Photon.Metrics.SmartReducer — Type.

Stores the calculated metrics. If multiple values are provided at the same step (like is the case with validation metrics), the moving average over those values will be stored.

source

Callbacks

Callbacks can be passed to the fit! function and provide a way to extend the functionality of Photon. Their functionality ranges from callbacks that generate output (like console or plots), to callbacks that save the model and callbacks that abort the training due to lack of progress.

Photon.Metrics.Meter — Type.

A meter is reponsible for presenting metric values. It can be used just as a regular callback argument to the fit! function. A meter is not limited to printing results to the console output, it can also be showing it on a TensorBoard or storing results in a database for example.

source

Photon.Metrics.ConsoleMeter — Type.

Logs metrics to the console output. By default it will only log at the end of an epoch and log the epoch, step, loss and validation loss. This is also the default configuration when you run fit! without specifying oter callbacks.

Usage

meter = ConsoleMeter([:loss, :val_accuracy, :accuracy]; epochOnly=false)
fit!(workout, data, val_data; epochs=5, cb=meter)

source

Photon.Metrics.SilentMeter — Type.

Use a SilentMeter in case no output is required. The default of fit! function is to use a ConsoleMeter and using the SilentMeter this behavior can be overriden.

Usage

fit!(workout, data, val_data; epochs=5, cb=SilentMeter())

source

Photon.Metrics.FileMeter — Type.

Logs metrics to a delimeted text file. By default it will log the metrics loss and val_loss at the end of each training step and end of the validation phase.

source

Photon.Metrics.PlotMeter — Type.

Plot metrics using the Plots module. At the end of each epoch the plot will be updated with the values of the metrics. This works especially nice if you are prototyping some model in an IDE like Juno.

source

Photon.Metrics.TensorBoardMeter — Type.

Logs metrics to a TensorBoard file so it can be viewed with TensorBoard. By default it will log the metrics loss and val_loss at the end of each training step and end of the validation phase.

This meter depends on the TensorBoardLogger to be installed. So if you didn't do so already, please run:

import Pkg; Pkg.add("TensorBoardLogger")

source

Photon.Metrics.EarlyStop — Type.

Stop the training if a certain metric didn't improve

source

Photon.Metrics.AutoSave — Type.

Save the Workout if a certain metric has improved since the last epoch

Usage

# save as long as the validation loss is declining
fit!(workout, data, cb=AutoSave(:val_loss))

source

Photon.Metrics.EpochSave — Type.

Save the Workout at the end of every epoch. Optionally provide a filename.

Usage

fit!(workout, data, cb=EpochSave())

source

Utils

Photon comes with a number of utility functions for common tasks with regards to using metrics.

Photon.Metrics.getmetricname — Function.

Function to generate the fully qualified metric name. It uses the metric name and the phase (:train or :valid) to come up with a unique name.

getmetricname(:loss, :train) # return is :loss
getmetricname(:loss, :valid) # return is :val_loss

source

Photon.Metrics.history — Function.

Get the history of a metric. The provided metric has the fully qualified name and the returned value is a tuple of steps and values.

h = history(workout, :val_loss)
# returns e.g ([1000, 2000, 3000, 4000], [0.81, 0.73, 0.64, 0.61])

h = history(workout, :loss)

source

Photon.plotmetrics — Function.

Plot the metrics after some training. This function will plot all the metrics in a single graph.

In order to avoid Photon being dependend on Plots, the calling code will have to provide that module as the first parameter.

Usage

fit!(workout, mydata, epochs=10)

import Plots
plotmetrics(Plots, workout)

source