Metrics

Metrics

Photon.Metrics.SmartReducer — Type.

Stores the calculated metrics

source

Photon.Metrics.BinaryAccuracy — Type.

Binary accuracy calculation

source

Photon.Metrics.OneHotBinaryAccuracy — Type.

Binary accuracy for a onehot classification.

source

Callbacks

Photon.Metrics.ConsoleMeter — Type.

Logs metrics to the console output. By default it will only log at the end of an epoch and log the epoch, step, loss and validation loss. This is also the default configuration when you run fit! without specifying your own meters.

Example:

meter = ConsoleMeter([:loss, :val_accuracy, :accuracy]; epochOnly=false)
fit!(workout, data, val_data; epochs=5, meters=[meter])

source

Photon.Metrics.FileMeter — Type.

Logs metrics to a delimeted text file. By default it will log the metrics loss and val_loss at the end of each training step and end of the validation phase.

source

Photon.Metrics.PlotMeter — Type.

Plot metrics using the Plots module. At the end of each epoch the plot will be updated with the values of the metrics. This works especially nice if you are prototyping some model in an IDE like Juno.

source

Photon.Metrics.TensorBoardMeter — Type.

Logs metrics to a TensorBoard file so it can be viewed with TensorBoard. By default it will log the metrics loss and val_loss at the end of each training step and end of the validation phase.

This meter depends on the TensorBoardLogger to be installed. So if you didn't do so already, please run:

import Pkg; Pkg.add("TensorBoardLogger")

source

Photon.Metrics.EarlyStop — Type.

Stop the training if a certain metric didn't improve

source

Photon.Metrics.AutoSave — Type.

Save the Workout if a certain metric has improved

Usage

# save as long as the validation loss is declining
fit!(workout, data, cb=AutoSave(:val_loss))

source

Photon.Metrics.EpochSave — Type.

Save the Workout at the end of every epoch.

Usage

fit!(workout, data, cb=EpochSave())

source

Utils

Photon.Metrics.history — Function.

Get the history of a metric. The provided metric has the fully qualified symbol and the returned value is a tuple of steps and values.

h = history(workout, :val_loss)
# returns e.g ([1000, 2000, 3000, 4000], [0.81, 0.73, 0.64, 0.61])

source